requirements.txt
# PyTorch with CUDA 12.1 (best for RTX 4090)
torch==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# MongoDB client
pymongo

# Hugging Face libraries
transformers
datasets
accelerate
peft
bitsandbytes
bert-score

# Jupyter support
ipykernel

# Progress bar utility
tqdm

# Key loading
python-dotenv
ipywidgets

# Data processing and manipulation
pandas
numpy
scikit-learn

#fronted
streamlit